[DESK-30] Create a Push CLI view under Repo Manager menu
[DESK-30] Create a Push CLI view under Repo Manager menu

Created a Push CLI view under Repo Manager menu
