+++ src/interface/CLI/views/repoManager/index.js
@@ -1,7 +1,9 @@
 const fetch = require('./fetch');
 const commit = require('./commit');
+const push = require('./push');
 
 module.exports = {
     fetch,
-    commit
+    commit,
+    push
 };
